# magalu/src/main/resources/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: Magalu Checkout & Payment API
  description: |
    **Reactive Checkout and Payment Platform**
    
    Professional checkout and payment backend developed with Java and Spring Boot, using reactive programming with Spring WebFlux and MySQL (R2DBC).
    
    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (RBAC)
    - Ownership validation (users can only access their own data)
    - Reactive non-blocking operations
    - RESTful API design
    
    ## Security
    Most endpoints require JWT Bearer token authentication. Use the `/api/auth/login` endpoint to obtain tokens.
    
    ## Roles
    - **CUSTOMER**: End customer with limited access to own data
    - **MERCHANT_ADMIN**: Store administrator with extended permissions
    - **SUPPORT**: Technical support with full access
    - **SYSTEM**: Internal system processes
  version: 1.0.0
  contact:
    name: Magalu Development Team
    email: dev@magalu.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.magalu.com
    description: Production server
...
paths:
...
  /api/auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Use a valid refresh token to obtain a new access token and refresh token pair.
        
        **Note:** The old refresh token will be revoked. Always use the new tokens returned in the response.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            examples:
              validRequest:
                summary: Valid refresh token request
                value:
                  refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  summary: New tokens issued
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.new_access_token...
                    refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.new_refresh_token...
                    expiresIn: 3600
                    tokenType: Bearer
        '401':
          description: Invalid, expired, or revoked refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidToken:
                  summary: Token is invalid or expired
                  value:
                    timestamp: "2026-02-09T10:30:00"
                    status: 401
                    error: Unauthorized
                    message: Refresh token expired or revoked
                    path: /api/auth/refresh-token
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.
        
        **How to use:**
        1. Call `/api/auth/login` to get tokens
        2. Use the `accessToken` in the Authorization header: `Authorization: Bearer <accessToken>`
        3. When access token expires, use `/api/auth/refresh-token` to get new tokens
  schemas:
...
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (expires in 5 minutes)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjEsImVtYWlsIjoidXNlckBleGFtcGxlLmNvbSIsInJvbGVzIjpbIkNVU1RPTUVSIl0sInR5cGUiOiJhY2Nlc3MiLCJpYXQiOjE2NDQzOTYwMDAsImV4cCI6MTY0NDM5NjMwMH0.signature
        refreshToken:
          type: string
          description: JWT refresh token (expires in 24 hours)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjEsImVtYWlsIjoidXNlckBleGFtcGxlLmNvbSIsInR5cGUiOiJyZWZyZXNoIiwiaWF0IjoxNjQ0Mzk2MDAwLCJleHAiOjE2NDQ0ODI0MDB9.signature
        expiresIn:
          type: integer
          format: int64
          description: Access token expiration time in seconds
          example: 3600
        tokenType:
          type: string
          description: Token type (always "Bearer")
          example: Bearer
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Valid refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2026-02-09T10:30:00"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: HTTP status text
          example: Bad Request
        errorCode:
          type: string
          description: Application-specific error code (optional)
          example: VALIDATION_ERROR
        message:
          type: string
          description: Error message
          example: Validation failed
        path:
          type: string
          description: Request path that generated the error
          example: /api/customers
        validationErrors:
          type: array
          description: List of validation errors (only present for 400 Bad Request)
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      properties:
        field:
          type: string
          description: Field name that failed validation
          example: email
        message:
          type: string
          description: Validation error message
          example: Email must be valid